# docs/openapi.yaml
openapi: 3.0.0
info:
  title: SGHSS API
  description: Sistema de Gestão Hospitalar e de Serviços de Saúde
  version: 1.0.0
  contact:
    name: Vinícius Pereira Carvalho
    email: contato@sghss.com

servers:
  - url: http://localhost:3000
    description: Servidor de desenvolvimento

tags:
  - name: Autenticação
    description: Endpoints de autenticação e autorização
  - name: Pacientes
    description: Gestão de pacientes
  - name: Consultas
    description: Agendamento e gestão de consultas

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        nome:
          type: string
        email:
          type: string
          format: email
        tipo:
          type: string
          enum: [PACIENTE, MEDICO, ENFERMEIRO, ADMINISTRADOR]
    
    Error:
      type: object
      properties:
        erro:
          type: string
        mensagem:
          type: string

paths:
  /api/auth/register:
    post:
      tags:
        - Autenticação
      summary: Registrar novo usuário
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - nome
                - email
                - senha
              properties:
                nome:
                  type: string
                  example: Maria Silva
                email:
                  type: string
                  format: email
                  example: maria@teste.com
                senha:
                  type: string
                  minLength: 6
                  example: senha123
                tipo:
                  type: string
                  enum: [PACIENTE, MEDICO, ENFERMEIRO, ADMINISTRADOR]
                  default: PACIENTE
      responses:
        '201':
          description: Usuário criado com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  mensagem:
                    type: string
                  usuario:
                    $ref: '#/components/schemas/User'
        '400':
          description: Campos obrigatórios não preenchidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Email já cadastrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  
  /api/auth/login:
    post:
      tags:
        - Autenticação
      summary: Fazer login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - senha
              properties:
                email:
                  type: string
                  format: email
                  example: maria@teste.com
                senha:
                  type: string
                  example: senha123
      responses:
        '200':
          description: Login realizado com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  mensagem:
                    type: string
                  token:
                    type: string
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        '401':
          description: Credenciais inválidas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  
  /api/auth/me:
    get:
      tags:
        - Autenticação
      summary: Obter dados do usuário logado
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Dados do usuário
          content:
            application/json:
              schema:
                type: object
                properties:
                  usuario:
                    $ref: '#/components/schemas/User'
        '401':
          description: Token não fornecido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  
  /api/pacientes:
    get:
      tags:
        - Pacientes
      summary: Listar pacientes
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de pacientes
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
        '401':
          description: Não autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'